if summon won combat the owner would lose 💀
i never really tested it too much so that just never happened

+++ Assets/scripts/MainGameScript/NewCombatSystem/AbilityManager.cs
@@ -334,6 +334,7 @@ public class AbilityManager : NetworkBehaviour
         {
             stats.isDead = true;
             stats.stats[Attributes.Health] = 0;
+            if(NewCombatManager.instance.allCombatants.Contains(this))
             NewCombatManager.instance.KILL(this);
         }
         onHit.Invoke();
